[MSHARED-1285] use an up-to-date scanner instead the newscanner (#77)
* [MSHARED-1285] - exhibit incremental issue

Current incremental build does not honor isUptodate(), so changed or
removed target resources are not refreshed until sources are modified or
a full build is performed.

2 new testcases exhibits this issue (one for missing target resource,
one for modified target resource). As stub BuildContext provides
appropriate isUptodate results, build should refresh this resources.

Current implementation prevents BuildContext implementors to trigger
appropriate resource refresh by tweaking isUptodate implementation.

See javadoc in BuildContext#newScanner javadoc that advices that
incremental build may be performed with a full resource scanning and a
isUptodate call to refresh only needed resources.

* MSHARED-1285 use an up-to-date scanner instead the newscanner

Currently it could happen that the scanner misses changed files (because
they are not part of the delta) or copies files even if they have not
changed (e.g. because the output has changes).

This uses now a different approach, instead of only handling the delta
files, we scan all inputs and compare if they are up-to-date with the
output.

---------

Co-authored-by: Laurent Almeras <laurent.almeras@kobalt.fr>
Co-authored-by: Christoph Läubrich <christoph@laeubi-soft.de>